fix cpp 06 ex01

+++ CPP06/ex01/includes/Serializer.hpp
@@ -6,7 +6,7 @@
 /*   By: jlorette <jlorette@42angouleme.fr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/26 15:39:46 by jlorette          #+#    #+#             */
-/*   Updated: 2025/04/14 12:44:30 by jlorette         ###   ########.fr       */
+/*   Updated: 2025/04/15 15:12:56 by jlorette         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,6 +14,9 @@
 #define SERIALIZER_HPP
 
 #include <Data.hpp>
+#include <iostream>
+#include <stdint.h>
+
 
 class Serializer
 {
@@ -24,8 +27,8 @@ class Serializer
 		Serializer& operator=(const Serializer& other);
 
 	public:
-		static unsigned long serialize(Data* ptr);
-		static Data* deserialize(unsigned long raw);
+		static uintptr_t serialize(Data *ptr);
+		static Data *deserialize(uintptr_t raw);
 };
 
 #endif
